Split in half to prove a point
